% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/organization.R
\name{organization}
\alias{organization}
\title{Get Information about a given organization}
\usage{
organization(organization_id, api_key)
}
\arguments{
\item{organization_id}{character vector of the organization_id. This id can be found in the URL of a organization profile.}

\item{api_key}{character vector of a valid API key.}
}
\value{
a data.frame containing basic information
}
\description{
This function returns organization information for a given organization_id
}

