% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/organization_id.R
\name{organization_id}
\alias{organization_id}
\title{Get organization_id for given organization profile url}
\usage{
organization_id(organization_url, api_key)
}
\arguments{
\item{organization_url}{character vector of the URL of organization_profile.}

\item{api_key}{character vector of a valid API key.}
}
\value{
The id of the organization.
}
\description{
This function returns the organization_id for a given
organization url. This id is needed to gather further
information about a petition.
}

