% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/organization_petitions.R
\name{organization_petitions}
\alias{organization_petitions}
\title{Get petitions created by a given organization}
\usage{
organization_petitions(petition_id, api_key)
}
\arguments{
\item{api_key}{character vector of a valid API key.}

\item{organization_id}{id as returned by \code{\link{organization_id}} or taken from organizations profile url}
}
\value{
A data.frame containing petitions including details.
}
\description{
This function returns the petitions a organization created
}
\seealso{
\code{\link{organization_id}}
}

