% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petition_id.R
\name{petition_id}
\alias{petition_id}
\title{Get petition_id for given url}
\usage{
petition_id(petition_url, api_key)
}
\arguments{
\item{petition_url}{character vector of the URL of a petition.}

\item{api_key}{character vector of a valid API key.}
}
\value{
The id of the petition.
}
\description{
This function returns the petition id for a given
petition url. This id is needed to gather further
information about a petition.
}

