% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petition_reasons.R
\name{petition_reasons}
\alias{petition_reasons}
\title{Get signature reasons for given petition id}
\usage{
petition_reasons(petition_id, nr = "all", sorting = "popularity",
  verbose = TRUE, api_key)
}
\arguments{
\item{petition_id}{petition id as returned by \code{\link{petition_id}}}

\item{nr}{number of reasons to be returned. Can be \code{numeric} or \code{"all"} to retrieve all.}

\item{sorting}{The order by which reasons will be returned. Accepted values are \code{"popularity"}, \code{"time_asc"}, and \code{"time_desc"}. Defaults to popularity (most number of likes)}

\item{verbose}{\code{logical} if console reporting is wanted. Defaults to \code{TRUE}}

\item{api_key}{character vector of a valid API key.}
}
\value{
A data.frame containing reasons including details.
}
\description{
This function returns the reasons for signing a given petition id.
}
\seealso{
\code{\link{petition_id}}
}

