% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petition_signatures.R
\name{petition_signatures}
\alias{petition_signatures}
\title{Get petition signatures for given petition id}
\usage{
petition_signatures(petition_id, nr = "all", sorting = "time_desc",
  verbose = TRUE, api_key)
}
\arguments{
\item{petition_id}{petition id as returned by \code{\link{petition_id}}}

\item{nr}{number of signatures to be returned. Can be \code{numeric} or \code{"all"} to retrieve all.}

\item{sorting}{The order by which signatures will be returned. Accepted values are \code{"time_asc"} and \code{"time_desc"}. If omitted, defaults to time_asc.}

\item{verbose}{\code{logical} if console reporting is wanted. Defaults to \code{TRUE}}

\item{api_key}{character vector of a valid API key.}
}
\value{
A data.frame containing signatures including details.
}
\description{
This function returns the signatures for a given petition id.  Note that retrieving all signatures may take quite some time since only 500 signatures can be gathered per request. Also signature count apparently does not necessarily match the number of signatures made availible via the API.
}
\seealso{
\code{\link{petition_id}}
}

