% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petition_target.R
\name{petition_targets}
\alias{petition_targets}
\title{Get petition targets for given petition id}
\usage{
petition_targets(petition_id, api_key)
}
\arguments{
\item{petition_id}{petition id as returned by \code{\link{petition_id}}:w}

\item{api_key}{character vector of a valid API key.}
}
\value{
A data.frame containing targets including details.
}
\description{
This function returns the target for a given petition id.
}
\seealso{
\code{\link{petition_id}}
}

