% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petition_updates.R
\name{petition_updates}
\alias{petition_updates}
\title{Get updates on petition}
\usage{
petition_updates(petition_id, nr = "all", sorting = c("time_asc",
  "time_desc"), verbose = TRUE, api_key)
}
\arguments{
\item{petition_id}{petition id as returned by \code{\link{petition_id}}}

\item{nr}{number of updates to be returned. Can be \code{numeric} or \code{"all"} to retrieve all.}

\item{sorting}{The order by which updates will be returned. Accepted values are \code{"time_asc"}, and \code{"time_desc"}.}

\item{verbose}{\code{logical} if console reporting is wanted. Defaults to \code{TRUE}}

\item{api_key}{character vector of a valid API key.}
}
\value{
A data.frame containing updates including details.
}
\description{
This function returns the updates for a given petition id.
}
\seealso{
\code{\link{petition_id}}
}

