% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user.R
\name{user}
\alias{user}
\title{Get Information about a given user}
\usage{
user(user_id, api_key)
}
\arguments{
\item{user_id}{character vector of the user_id. This id can be found in the URL of a user profile.}

\item{api_key}{character vector of a valid API key.}
}
\value{
a data.frame containing basic information
}
\description{
This function returns user information for a given user_id
}

