% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user_id.R
\name{user_id}
\alias{user_id}
\title{Get user_id for given userpfrofile url}
\usage{
user_id(user_url, api_key)
}
\arguments{
\item{user_url}{character vector of the URL of user_profile.}

\item{api_key}{character vector of a valid API key.}
}
\value{
The id of the user.
}
\description{
This function returns the user_id for a given
petition url. This id is needed to gather further
information about a petition.
}

