% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/user_petitions.R
\name{user_petitions}
\alias{user_petitions}
\title{Get petitions created by a given user}
\usage{
user_petitions(petition_id, api_key)
}
\arguments{
\item{api_key}{character vector of a valid API key.}

\item{user_id}{id as returned by \code{\link{user_id}} or taken from userprofile url}
}
\value{
A data.frame containing petitions including details.
}
\description{
This function returns the petitions a user created
}
\seealso{
\code{\link{user_id}}
}

